Fixing type and number problem primary. Changing a bit and formatting. Resolving the number and type problem. Creating `dllst.py` for testing single function of dll. Removing unneseccary files. Updating unit_test.py to use the right data type else bad things happen. Updating test.csv with the needed type. Updating library.cpp

+++ src/basics/factorial.cpp
@@ -1,7 +1,9 @@
+#include <cstdint>
+
 class moo
 {
 public:
-    static int factorial(int n)
+    static int64_t factorial(int n)
     {
         if (n <= 1) return 1;
         return n * factorial(n - 1);
@@ -10,7 +12,7 @@ public:
 
 // @formatter:off
 extern "C" {
-    __declspec(dllexport) int fac(int n)
+    __declspec(dllexport) int64_t fac(int n)
     {
         return moo::factorial(n);
     }
